packages/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# PostgreSQL
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
greenlet==3.0.3

# MongoDB
motor==3.6.0
pymongo==4.9.0
beanie==1.26.0

# Redis
redis==5.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# WebSocket
python-socketio==5.11.0

# Utilities
httpx==0.26.0
